was wondering why ever n after 34 was returning 0..The factorial wasn't a float

+++ src/Assignment02/Part1Main.java
@@ -38,7 +38,7 @@ public class Part1Main {
         System.out.println("The factorial of " + n + " is: " + factorial(n));
     }
 
-    public static int factorial(int n) {
+    public static float factorial(int n) {
         if (n == 0) {
             return 1;
         } else {
